+++ packages/euclid/src/msgs/vlp.rs
@@ -2,7 +2,7 @@ use crate::{
     chain::{ChainUid, CrossChainUser},
     fee::{Fee, TotalFees},
     swap::NextSwapVlp,
-    token::{Pair, PairWithAmount, Token, TokenType},
+    token::{Pair, PairWithAmount, Token},
 };
 use cosmwasm_schema::{cw_serde, QueryResponses};
 use cosmwasm_std::Uint128;
@@ -162,16 +162,6 @@ pub struct AllPoolsResponse {
 
 #[cw_serde]
 pub struct MigrateMsg {}
-
-#[cw_serde]
-pub struct VlpRemoveLiquidityResponse {
-    pub liquidity_released: PairWithAmount,
-    pub preferred_denom: Option<TokenType>,
-    pub burn_lp_tokens: Uint128,
-    pub tx_id: String,
-    pub sender: CrossChainUser,
-    pub vlp_address: String,
-}
 
 #[cw_serde]
 pub struct VlpSwapResponse {
